Transfer potentially inactive GitHub account?
Hello,

I am part of a research group in Trondheim, Norway, working on developing solutions for improving cancer diagnostics and patient care.

Our research group has taken the name AICAN (https://www.ntnu.edu/ikom/aican#/view/about), and it was of interest to create a GitHub user as main channel to share all our source code and research, as we promote open research and solutions that benefit patient care. Having a username that matches our research group name would be ideal.

However, this username is already taken, as you already know ;-), but it seems like your account may be inactive. If the account indeed is inactive, or if the username is not that important to you, would you be able to transfer the account name to us? That would be much obliged!🥇

If you are interested, or if not, please, reply to my e-mail:
andrped94@gmail.com

Thank you, for taking your time to read this 👍

Cheers from Norway,
André Pedersen